Mem api (#259)
* Start memory API tests using RAM buffer

* Add write_read_i16 test

* Add write_read_float

* Add test write_read_float_array

* Add test write_read_u32_array

* Rename offset -> reg

* Add run-time versions tests

* Rename MemoryMap -> Memory

* Add test set_get_ptr_float

* Add test set_get_ptr_u32

* Add run-time versions

* Update blink

* Fix line error in flask-uwsgi.ini

+++ scripts/templates/memory.hpp
@@ -13,7 +13,7 @@ extern "C" {
 
 namespace mem {
 {% for addr in dic['memory'] -%}
-constexpr uint32_t {{ addr['name'] }} = {{ loop.index0 }};
+constexpr size_t {{ addr['name'] }} = {{ loop.index0 }};
 constexpr uintptr_t {{ addr['name'] }}_addr = {{ addr['offset'] }};
 constexpr uint32_t {{ addr['name'] }}_range = {{ addr['range'] | replace_KMG }};
 constexpr uint32_t {{ addr['name'] }}_nblocks = {{ addr['n_blocks'] }};
